moved morphic functionality into client/morphic folder
