+++ draft/test.html
@@ -87,14 +87,19 @@
     }).catch(function(err) { alert("load Lively4 failed")});
 
     var morphic;
-    System.import(lively4url + "/src/client/morphic.js").then(m => {
+    System.import(lively4url + "/src/client/morphic/main.js").then(m => {
         morphic = m;
         morphic.initMorphicTools();
-    }).catch(function(err) { debugger; alert("load morphic failed"); });
+    }).catch(function(err) {
+        debugger; alert("load morphic/main failed");
+    });
 
-    System.import(lively4url + "/src/client/morphic-toolbox.js").then(toolbox => {
+    System.import(lively4url + "/src/client/morphic/toolbox.js").then(toolbox => {
         document.body.appendChild(toolbox.createMorphicToolbox());
-    }).catch(function(err) { debugger; alert("load morphic-toolbox failed"); });
+    }).catch(function(err) {
+        debugger; alert("load morphic/toolbox failed");
+    });
+
 </script>
 </body>
 </html>
